requirements.txt
# absl-py==1.0.0
# cachetools==5.0.0
# certifi==2021.10.8
# charset-normalizer==2.0.12
# colorama==0.4.4
# cycler==0.11.0
# easyocr==1.4.2
# fonttools==4.33.3
# google-auth==2.6.6
# google-auth-oauthlib==0.4.6
# grpcio==1.46.0
# idna==3.3
# imageio==2.19.0
# importlib-metadata==4.11.3
# kiwisolver==1.4.2
# Markdown==3.3.7
# networkx==2.8
# numpy==1.26.4
# oauthlib==3.2.0
# opencv-contrib-python==4.5.5.64
# opencv-python==4.3.0.38
# packaging==21.3
# pandas==2.2.3
# Pillow==9.1.0
# protobuf==3.20.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pyparsing==3.0.8
# python-bidi==0.4.2
# python-dateutil==2.8.2
# pytz==2022.1
# PyWavelets==1.3.0
# PyYAML==6.0
# requests==2.27.1
# requests-oauthlib==1.3.1
# rsa==4.8
# scikit-image==0.19.2
# scipy==1.8.0
# six==1.16.0
# tensorboard==2.9.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tifffile==2022.5.4
# torch==1.11.0+cu113
# torchaudio==0.11.0+cu113
# torchvision==0.12.0+cu113
# tqdm==4.64.0
# typing_extensions==4.2.0
# urllib3==1.26.9
# Werkzeug==2.1.2
# wincertstore==0.2
# zipp==3.8.0
XlsxWriter==3.2.3
streamlit==1.44.1
easyocr==1.4.2
pillow==10.4.0
scikit-learn==1.5.2
scipy==1.13.1
torch==1.13.1
torchvision==0.14.1
ultralytics==8.2.34
pandas==2.2.3
opencv-python>=4.6.0
PyYAML==6.0.2
tqdm==4.67.1
onnx==1.16.0
onnxruntime==1.17.0
onnxruntime-gpu==1.17.0
tensorrt==8.6.1






